moved arguments to the tools file
