+++ tools.py
@@ -1,4 +1,6 @@
 from pyfiglet import Figlet
+import argparse as ap
+import sys
 
 def blue(text: str) -> str:
     """
@@ -30,6 +32,13 @@ def red(text: str) -> str:
     """
     return "\033[31m" + text + "\033[0m"
 
+def clear_screen():
+    """
+    It prints 25 new lines
+    """
+    print("\n" * 25)
+
+
 def show_splash():
     """
     Display splash screen
@@ -38,3 +47,40 @@ def show_splash():
     title = "R3ddit\n Scrapper"
     f = Figlet(font="standard")
     print(blue(f.renderText(title)))
+
+
+def argument():
+    """Parsing the arguments passed by the user.
+    optional arguments:
+    -s --sub: The subreddit you want to download from
+    -l --limit: The number of images to download
+    -o --order: hot, top, new
+    -n --nsfw: Set to False of you want to download all NON NSFW images
+    """
+    parser = ap.ArgumentParser(description="R3ddit Scrapper")
+    parser.add_argument(
+        "-s", "--sub", help="The subreddit you want to download from", required=False
+    )
+    parser.add_argument(
+        "-l", "--limit", help="The number of images to download", required=False
+    )
+    parser.add_argument("-o", "--order", help="hot, top, new", required=False)
+    parser.add_argument(
+        "-n",
+        "--nsfw",
+        help="Set to False of you want to download all NON NSFW images",
+        required=False,
+    )
+    # if no arguments are passed, return
+    if len(sys.argv) == 1:
+        return
+    ol = ["hot", "top", "new"]
+    sub = parser.parse_args().sub if parser.parse_args().sub else "pics"
+    limit = int(parser.parse_args().limit) if parser.parse_args().limit else 1
+    order = parser.parse_args().order if parser.parse_args().order in ol else "hot"
+    nsfw = parser.parse_args().nsfw if parser.parse_args().nsfw else "True"
+    print(f"Subreddit: {sub}")
+    print(f"Limit: {limit}")
+    print(f"Order: {order}")
+    print(f"NSFW: {nsfw}")
+    return sub, limit, order, nsfw
